[runtime] Split JSON parser and serializer into separate modules
