+++ src/js/runtime/intrinsics/json_object.rs
@@ -1,33 +1,29 @@
 use std::collections::HashSet;
 
 use crate::{
-    common::{unicode::is_surrogate_code_point, wtf_8::Wtf8String},
-    must, must_a,
+    common::wtf_8::Wtf8String,
+    must,
     runtime::{
         abstract_operations::{
-            call, call_object, create_data_property, create_data_property_or_throw,
-            enumerable_own_property_names, get_v, length_of_array_like, KeyOrValue,
+            call_object, create_data_property, create_data_property_or_throw,
+            enumerable_own_property_names, length_of_array_like, KeyOrValue,
         },
         alloc_error::AllocResult,
-        array_object::array_create,
-        error::{syntax_error, type_error},
+        error::syntax_error,
         function::get_argument,
         get,
         intrinsics::rust_runtime::RuntimeFunction,
         object_value::ObjectValue,
         ordinary_object::ordinary_object_create,
         property::Property,
-        string_parsing::{parse_signed_decimal_literal, StringLexer},
-        string_value::StringValue,
+        string_parsing::StringLexer,
         to_string,
-        type_utilities::{
-            is_array, is_callable, number_to_string, to_integer_or_infinity, to_number,
-        },
+        type_utilities::{is_array, is_callable, to_integer_or_infinity, to_number},
         Context, EvalResult, Handle, PropertyKey, Realm, Value,
     },
 };
 
-use super::intrinsics::Intrinsic;
+use super::{intrinsics::Intrinsic, json_parser::parse_json, json_serializer::JSONSerializer};
 
 /// The JSON Object (https://tc39.es/ecma262/#sec-json-object)
 pub struct JSONObject;
@@ -256,554 +252,5 @@ impl JSONObject {
         }
 
         Ok(serializer.build(cx)?.as_value())
-    }
-}
-
-enum JSONValue {
-    Object(Vec<(Wtf8String, JSONValue)>),
-    Array(Vec<JSONValue>),
-    String(Wtf8String),
-    Number(f64),
-    Boolean(bool),
-    Null,
-}
-
-fn parse_json(lexer: &mut StringLexer) -> Option<JSONValue> {
-    // Could have leading whitespace
-    skip_json_whitespace(lexer);
-
-    let value = parse_json_value(lexer)?;
-
-    // Could have trailing whitespace
-    skip_json_whitespace(lexer);
-
-    // Must be end of input otherwise is invalid JSON
-    if !lexer.is_end() {
-        return None;
-    }
-
-    Some(value)
-}
-
-fn parse_json_value(lexer: &mut StringLexer) -> Option<JSONValue> {
-    if lexer.eat('{') {
-        parse_json_object(lexer)
-    } else if lexer.eat('[') {
-        parse_json_array(lexer)
-    } else if lexer.eat('"') {
-        let string_value = parse_json_string(lexer)?;
-        Some(JSONValue::String(string_value))
-    } else if lexer.eat('t') {
-        lexer.expect('r')?;
-        lexer.expect('u')?;
-        lexer.expect('e')?;
-
-        Some(JSONValue::Boolean(true))
-    } else if lexer.eat('f') {
-        lexer.expect('a')?;
-        lexer.expect('l')?;
-        lexer.expect('s')?;
-        lexer.expect('e')?;
-
-        Some(JSONValue::Boolean(false))
-    } else if lexer.eat('n') {
-        lexer.expect('u')?;
-        lexer.expect('l')?;
-        lexer.expect('l')?;
-
-        Some(JSONValue::Null)
-    } else if lexer.current_is_decimal_digit()
-        || lexer.current_equals('-')
-        || lexer.current_equals('+')
-    {
-        // JSON numbers do not allow leading zeros
-        if lexer.current_equals('0') && lexer.peek_ascii_char() == Some('0') {
-            return None;
-        }
-
-        let number = parse_signed_decimal_literal(lexer)?;
-        Some(JSONValue::Number(number))
-    } else {
-        None
-    }
-}
-
-fn skip_json_whitespace(lexer: &mut StringLexer) {
-    while lexer.current_is_json_whitespace() {
-        lexer.advance();
-    }
-}
-
-fn parse_json_object(lexer: &mut StringLexer) -> Option<JSONValue> {
-    let mut properties = vec![];
-
-    // Consume leading whitespace before the next character (key or closing brace)
-    skip_json_whitespace(lexer);
-
-    while !lexer.current_equals('}') && !lexer.is_end() {
-        // Parse the string key
-        lexer.expect('"')?;
-        let key_string = parse_json_string(lexer)?;
-
-        // Key separated by value by colon
-        skip_json_whitespace(lexer);
-        lexer.expect(':')?;
-
-        // Parse the value
-        skip_json_whitespace(lexer);
-        let value = parse_json_value(lexer)?;
-
-        // Add the key value pair to the object
-        properties.push((key_string, value));
-
-        // There may be whitespace after the property and before the comma/closing brace
-        skip_json_whitespace(lexer);
-
-        if !lexer.eat(',') {
-            break;
-        }
-
-        // There may be whitespace after the comma and before the next property
-        skip_json_whitespace(lexer);
-
-        // Trailing comma is not allowed
-        if lexer.current_equals('}') {
-            return None;
-        }
-    }
-
-    lexer.expect('}')?;
-
-    Some(JSONValue::Object(properties))
-}
-
-fn parse_json_array(lexer: &mut StringLexer) -> Option<JSONValue> {
-    let mut elements = vec![];
-
-    // Consume leading whitespace before the next character (value or closing bracket)
-    skip_json_whitespace(lexer);
-
-    while !lexer.current_equals(']') && !lexer.is_end() {
-        // Parse the array element
-        let element = parse_json_value(lexer)?;
-        elements.push(element);
-
-        // There may be whitespace after a value and before the comma/closing bracket
-        skip_json_whitespace(lexer);
-
-        if !lexer.eat(',') {
-            break;
-        }
-
-        // There may be whitespace after the comma and before the next value
-        skip_json_whitespace(lexer);
-
-        // Trailing comma is not allowed
-        if lexer.current_equals(']') {
-            return None;
-        }
-    }
-
-    lexer.expect(']')?;
-
-    Some(JSONValue::Array(elements))
-}
-
-fn parse_json_string(lexer: &mut StringLexer) -> Option<Wtf8String> {
-    let mut string = Wtf8String::new();
-
-    while !lexer.current_equals('"') && !lexer.is_end() {
-        // Escape sequence which must be decoded
-        if lexer.eat('\\') {
-            if lexer.current_equals('"') {
-                string.push_char('"');
-                lexer.advance();
-            } else if lexer.current_equals('\\') {
-                string.push_char('\\');
-                lexer.advance();
-            } else if lexer.current_equals('/') {
-                string.push_char('/');
-                lexer.advance();
-            } else if lexer.current_equals('b') {
-                string.push_char('\x08');
-                lexer.advance();
-            } else if lexer.current_equals('f') {
-                string.push_char('\x0C');
-                lexer.advance();
-            } else if lexer.current_equals('n') {
-                string.push_char('\n');
-                lexer.advance();
-            } else if lexer.current_equals('r') {
-                string.push_char('\r');
-                lexer.advance();
-            } else if lexer.current_equals('t') {
-                string.push_char('\t');
-                lexer.advance();
-            } else if lexer.eat('u') {
-                let mut code_point = 0;
-
-                for _ in 0..4 {
-                    code_point <<= 4;
-                    code_point += lexer.current_hex_value()?;
-
-                    lexer.advance();
-                }
-
-                // Control characters are not allowed
-                if code_point <= 0x1F {
-                    return None;
-                }
-
-                string.push(code_point)
-            } else {
-                return None;
-            }
-        } else if let Some(code_unit) = lexer.current() {
-            // Control characters are not allowed
-            if code_unit <= 0x1F {
-                return None;
-            }
-
-            // Otherwise is a regular code unit which can be directly added to string
-            string.push(code_unit as u32);
-            lexer.advance();
-        } else {
-            return None;
-        }
-    }
-
-    lexer.expect('"')?;
-
-    Some(string)
-}
-
-impl JSONValue {
-    pub fn to_js_value(&self, mut cx: Context) -> EvalResult<Handle<Value>> {
-        let value = match self {
-            Self::Null => cx.null(),
-            Self::Boolean(b) => cx.bool(*b),
-            Self::Number(n) => Value::from(*n).to_handle(cx),
-            Self::String(s) => cx.alloc_wtf8_string(s)?.into(),
-            Self::Array(values) => {
-                let array = must_a!(array_create(cx, 0, None));
-
-                // Key is shared between iterations
-                let mut key = PropertyKey::uninit().to_handle(cx);
-
-                for (i, value) in values.iter().enumerate() {
-                    key.replace(PropertyKey::from_u64(cx, i as u64)?);
-                    let desc = Property::data(value.to_js_value(cx)?, true, true, true);
-                    array.as_object().set_property(cx, key, desc)?;
-                }
-
-                array.into()
-            }
-            Self::Object(properties) => {
-                let object = ordinary_object_create(cx)?;
-
-                // Key is shared between iterations
-                let mut key_value: Handle<Value> = Handle::empty(cx);
-                let mut key = PropertyKey::uninit().to_handle(cx);
-
-                for (key_string, value) in properties {
-                    key_value.replace(cx.alloc_wtf8_string_ptr(key_string)?.as_string().into());
-                    key.replace(PropertyKey::string(cx, key_value.as_string())?);
-
-                    let value = value.to_js_value(cx)?;
-
-                    must_a!(create_data_property_or_throw(cx, object, key, value));
-                }
-
-                object.into()
-            }
-        };
-
-        Ok(value)
-    }
-}
-
-struct JSONSerializer {
-    /// String that we are building
-    builder: Wtf8String,
-
-    stack: Vec<Handle<ObjectValue>>,
-    replacer_function: Option<Handle<ObjectValue>>,
-    property_list: Option<Vec<Handle<PropertyKey>>>,
-    indent: u32,
-    gap: Wtf8String,
-}
-
-impl JSONSerializer {
-    fn new(
-        replacer_function: Option<Handle<ObjectValue>>,
-        property_list: Option<Vec<Handle<PropertyKey>>>,
-        gap: Wtf8String,
-    ) -> Self {
-        JSONSerializer {
-            builder: Wtf8String::new(),
-            stack: vec![],
-            replacer_function,
-            property_list,
-            indent: 0,
-            gap,
-        }
-    }
-
-    fn build(self, mut cx: Context) -> EvalResult<Handle<StringValue>> {
-        Ok(cx.alloc_wtf8_string(&self.builder)?.as_string())
-    }
-
-    /// SerializeJSONProperty (https://tc39.es/ecma262/#sec-serializejsonproperty)
-    ///
-    /// Returns whether the property was successfully serialized. False is equivalent to returning
-    /// undefined in the spec.
-    fn serialize_json_property(
-        &mut self,
-        cx: Context,
-        key: Handle<PropertyKey>,
-        holder: Handle<ObjectValue>,
-    ) -> EvalResult<bool> {
-        let mut value = get(cx, holder, key)?;
-
-        // Call toJSON method if one is present
-        if value.is_object() || value.is_bigint() {
-            let to_json = get_v(cx, value, cx.names.to_json())?;
-            if is_callable(to_json) {
-                let key_value = key.to_value(cx)?;
-                value = call(cx, to_json, value, &[key_value])?;
-            }
-        }
-
-        // Call replacer function if one was provided
-        if let Some(replacer_function) = self.replacer_function {
-            let key_value = key.to_value(cx)?;
-            value = call_object(cx, replacer_function, holder.into(), &[key_value, value])?;
-        }
-
-        // Convert primitive wrapper objects to their underlying primitive value
-        if value.is_object() {
-            let value_object = value.as_object();
-            if value_object.is_number_object() {
-                value = to_number(cx, value)?;
-            } else if value_object.is_string_object() {
-                value = to_string(cx, value)?.into();
-            } else if let Some(boolean_object) = value_object.as_boolean_object() {
-                value = cx.bool(boolean_object.boolean_data());
-            } else if let Some(bigint_object) = value_object.as_bigint_object() {
-                value = bigint_object.bigint_data().into()
-            }
-        }
-
-        if value.is_null() {
-            self.builder.push_str("null");
-        } else if value.is_bool() {
-            if value.as_bool() {
-                self.builder.push_str("true");
-            } else {
-                self.builder.push_str("false");
-            }
-        } else if value.is_string() {
-            self.serialize_json_string(value.as_string())?;
-        } else if value.is_number() {
-            let number = value.as_number();
-            if number.is_finite() {
-                // Inlined ToString on number
-                self.builder.push_str(&number_to_string(number));
-            } else {
-                self.builder.push_str("null");
-            }
-        } else if value.is_bigint() {
-            return type_error(cx, "BigInt value cannot be serialized to JSON");
-        } else if value.is_object() && !is_callable(value) {
-            if is_array(cx, value)? {
-                self.serialize_json_array(cx, value.as_object())?;
-            } else {
-                self.serialize_json_object(cx, value.as_object())?;
-            }
-        } else {
-            return Ok(false);
-        }
-
-        Ok(true)
-    }
-
-    /// QuoteJSONString (https://tc39.es/ecma262/#sec-quotejsonstring)
-    fn serialize_json_string(&mut self, string: Handle<StringValue>) -> AllocResult<()> {
-        self.builder.push_char('"');
-
-        for code_point in string.iter_code_points()? {
-            // Escape characters
-            if code_point == '\x08' as u32 {
-                self.builder.push_str("\\b")
-            } else if code_point == '\x0C' as u32 {
-                self.builder.push_str("\\f")
-            } else if code_point == '\t' as u32 {
-                self.builder.push_str("\\t")
-            } else if code_point == '\n' as u32 {
-                self.builder.push_str("\\n")
-            } else if code_point == '\r' as u32 {
-                self.builder.push_str("\\r")
-            } else if code_point == '\"' as u32 {
-                self.builder.push_str("\\\"")
-            } else if code_point == '\\' as u32 {
-                self.builder.push_str("\\\\")
-            } else if code_point < 0x20 || is_surrogate_code_point(code_point) {
-                // Must be serialized as a unicode escape
-                self.builder.push_str(&format!("\\u{code_point:04x}"))
-            } else {
-                self.builder.push(code_point);
-            }
-        }
-
-        self.builder.push_char('"');
-
-        Ok(())
-    }
-
-    fn indent(&mut self) {
-        for _ in 0..self.indent {
-            self.builder.push_wtf8_str(&self.gap);
-        }
-    }
-
-    fn check_for_cycle(&mut self, cx: Context, value: Handle<ObjectValue>) -> EvalResult<()> {
-        let value_ptr = *value;
-        let has_cycle = self
-            .stack
-            .iter()
-            .any(|ancestor| ancestor.ptr_eq(&value_ptr));
-
-        if has_cycle {
-            type_error(cx, "cyclic object cannot be serialized to JSON")
-        } else {
-            Ok(())
-        }
-    }
-
-    /// SerializeJSONObject (https://tc39.es/ecma262/#sec-serializejsonobject)
-    fn serialize_json_object(
-        &mut self,
-        cx: Context,
-        object: Handle<ObjectValue>,
-    ) -> EvalResult<()> {
-        self.check_for_cycle(cx, object)?;
-
-        let mut keys = vec![];
-        if let Some(property_list) = &self.property_list {
-            keys.extend_from_slice(property_list);
-        } else {
-            let property_key_values = enumerable_own_property_names(cx, object, KeyOrValue::Key)?;
-            for property_key_value in property_key_values {
-                let property_key = PropertyKey::from_value(cx, property_key_value)?;
-                keys.push(property_key.to_handle(cx));
-            }
-        };
-
-        self.builder.push_char('{');
-
-        self.stack.push(object);
-        self.indent += 1;
-
-        let mut has_property = false;
-
-        for key in keys {
-            let string_key = key.to_value(cx)?;
-            debug_assert!(string_key.is_string());
-
-            // Mark the string position to revert to if this property should not be added
-            let undo_length = self.builder.len();
-
-            // Insert a comma and indentations before all properties but the first
-            if has_property {
-                if self.gap.is_empty() {
-                    self.builder.push_char(',');
-                } else {
-                    self.builder.push_str(",\n");
-                    self.indent();
-                }
-            } else if !self.gap.is_empty() {
-                // There may be a newline before the first property
-                self.builder.push_char('\n');
-                self.indent();
-            }
-
-            // Serialize the key and value separated by a colon
-            self.serialize_json_string(string_key.as_string())?;
-
-            self.builder.push_char(':');
-            if !self.gap.is_empty() {
-                self.builder.push_char(' ');
-            }
-
-            if self.serialize_json_property(cx, key, object)? {
-                has_property = true;
-            } else {
-                // If we cannot serialize property, undo back to before this property was added
-                self.builder.truncate(undo_length);
-            }
-        }
-
-        self.indent -= 1;
-        self.stack.pop();
-
-        // There may be a newline after the last property
-        if has_property && !self.gap.is_empty() {
-            self.builder.push_char('\n');
-            self.indent();
-        }
-
-        self.builder.push_char('}');
-
-        Ok(())
-    }
-
-    /// SerializeJSONArray (https://tc39.es/ecma262/#sec-serializejsonarray)
-    fn serialize_json_array(&mut self, cx: Context, array: Handle<ObjectValue>) -> EvalResult<()> {
-        self.check_for_cycle(cx, array)?;
-
-        self.builder.push_char('[');
-
-        self.stack.push(array);
-        self.indent += 1;
-
-        // Share key between iterations
-        let mut key = PropertyKey::uninit().to_handle(cx);
-
-        let length = length_of_array_like(cx, array)?;
-        for i in 0..length {
-            key.replace(PropertyKey::from_u64(cx, i)?);
-
-            // Insert a comma and indentations before all elements but the first
-            if i != 0 {
-                if self.gap.is_empty() {
-                    self.builder.push_char(',');
-                } else {
-                    self.builder.push_str(",\n");
-                    self.indent();
-                }
-            } else if !self.gap.is_empty() {
-                // There may be a newline before the first element
-                self.builder.push_char('\n');
-                self.indent();
-            }
-
-            // Serialize each element, replacing with null if it cannot be serialized
-            if !self.serialize_json_property(cx, key, array)? {
-                self.builder.push_str("null");
-            }
-        }
-
-        self.indent -= 1;
-        self.stack.pop();
-
-        // There may be a newline after the last element
-        if length != 0 && !self.gap.is_empty() {
-            self.builder.push_char('\n');
-            self.indent();
-        }
-
-        self.builder.push_char(']');
-
-        Ok(())
     }
 }
